Fixed errors.

+++ gl/shader/shader_program.go
@@ -217,9 +217,9 @@ func (p *ShaderProgram) SetTexture(name string, texture_unit int, texture_handle
 		return -1
 	}
 
-	wrapper.ActiveTexture(gl.TEXTURE0 + texture_unit)
+	wrapper.ActiveTexture(uint32(gl.TEXTURE0 + texture_unit))
 	texture.BindTexture(texture_handle)
-	wrapper.Uniform1i(location, texture_unit)
+	wrapper.Uniform1i(location, int32(texture_unit))
 	texture.UnbindTexture()
 
 	return 0
